refactor: Renomme le dossier
