+++ src/client/components/features/Formation/FormulaireRecherche/FormulaireRechercherFormation.tsx
@@ -6,7 +6,7 @@ import styles
 import { ButtonComponent } from '~/client/components/ui/Button/ButtonComponent';
 import {
 	ComboboxMetiers,
-} from '~/client/components/ui/Form/Combobox/InputAutocomplétionMétier/ComboboxMetiers';
+} from '~/client/components/ui/Form/Combobox/ComboboxMetiers/ComboboxMetiers';
 import { InputCommune } from '~/client/components/ui/Form/InputCommune/InputCommune';
 import { Icon } from '~/client/components/ui/Icon/Icon';
 import { Select } from '~/client/components/ui/Select/Select';
